Merge upstream changes

+++ internal/entity/face.go
@@ -9,11 +9,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/jinzhu/gorm"
-
 	"github.com/photoprism/photoprism/internal/face"
-
-	"github.com/photoprism/photoprism/pkg/clusters"
+	"github.com/photoprism/photoprism/pkg/rnd"
 )
 
 var faceMutex = sync.Mutex{}
@@ -29,7 +26,7 @@ type Face struct {
 	Collisions      int             `json:"Collisions" yaml:"Collisions,omitempty"`
 	CollisionRadius float64         `json:"CollisionRadius" yaml:"CollisionRadius,omitempty"`
 	EmbeddingJSON   json.RawMessage `gorm:"type:MEDIUMBLOB;" json:"-" yaml:"EmbeddingJSON,omitempty"`
-	embedding       Embedding       `gorm:"-"`
+	embedding       face.Embedding  `gorm:"-"`
 	MatchedAt       *time.Time      `json:"MatchedAt" yaml:"MatchedAt,omitempty"`
 	CreatedAt       time.Time       `json:"CreatedAt" yaml:"CreatedAt,omitempty"`
 	UpdatedAt       time.Time       `json:"UpdatedAt" yaml:"UpdatedAt,omitempty"`
@@ -44,7 +41,7 @@ func (Face) TableName() string {
 }
 
 // NewFace returns a new face.
-func NewFace(subjUID, faceSrc string, embeddings Embeddings) *Face {
+func NewFace(subjUID, faceSrc string, embeddings face.Embeddings) *Face {
 	result := &Face{
 		SubjUID: subjUID,
 		FaceSrc: faceSrc,
@@ -58,8 +55,8 @@ func NewFace(subjUID, faceSrc string, embeddings Embeddings) *Face {
 }
 
 // SetEmbeddings assigns face embeddings.
-func (m *Face) SetEmbeddings(embeddings Embeddings) (err error) {
-	m.embedding, m.SampleRadius, m.Samples = EmbeddingsMidpoint(embeddings)
+func (m *Face) SetEmbeddings(embeddings face.Embeddings) (err error) {
+	m.embedding, m.SampleRadius, m.Samples = face.EmbeddingsMidpoint(embeddings)
 
 	// Limit sample radius to reduce false positives.
 	if m.SampleRadius > 0.35 {
@@ -93,9 +90,9 @@ func (m *Face) Matched() error {
 }
 
 // Embedding returns parsed face embedding.
-func (m *Face) Embedding() Embedding {
+func (m *Face) Embedding() face.Embedding {
 	if len(m.EmbeddingJSON) == 0 {
-		return Embedding{}
+		return face.Embedding{}
 	} else if len(m.embedding) > 0 {
 		return m.embedding
 	} else if err := json.Unmarshal(m.EmbeddingJSON, &m.embedding); err != nil {
@@ -106,10 +103,10 @@ func (m *Face) Embedding() Embedding {
 }
 
 // Match tests if embeddings match this face.
-func (m *Face) Match(embeddings Embeddings) (match bool, dist float64) {
+func (m *Face) Match(embeddings face.Embeddings) (match bool, dist float64) {
 	dist = -1
 
-	if len(embeddings) == 0 {
+	if embeddings.Empty() {
 		// Np embeddings, no match.
 		return false, dist
 	}
@@ -121,9 +118,9 @@ func (m *Face) Match(embeddings Embeddings) (match bool, dist float64) {
 		return false, dist
 	}
 
-	// Calculate smallest distance to embeddings.
+	// Calculate the smallest distance to embeddings.
 	for _, e := range embeddings {
-		if d := clusters.EuclideanDistance(e, faceEmbedding); d < dist || dist < 0 {
+		if d := e.Distance(faceEmbedding); d < dist || dist < 0 {
 			dist = d
 		}
 	}
@@ -133,7 +130,7 @@ func (m *Face) Match(embeddings Embeddings) (match bool, dist float64) {
 	case dist < 0:
 		// Should never happen.
 		return false, dist
-	case dist > (m.SampleRadius + face.MatchRadius):
+	case dist > (m.SampleRadius + face.MatchDist):
 		// Too far.
 		return false, dist
 	case m.CollisionRadius > 0.1 && dist > m.CollisionRadius:
@@ -146,7 +143,7 @@ func (m *Face) Match(embeddings Embeddings) (match bool, dist float64) {
 }
 
 // ResolveCollision resolves a collision with a different subject's face.
-func (m *Face) ResolveCollision(embeddings Embeddings) (resolved bool, err error) {
+func (m *Face) ResolveCollision(embeddings face.Embeddings) (resolved bool, err error) {
 	if m.SubjUID == "" {
 		// Ignore reports for anonymous faces.
 		return false, nil
@@ -258,7 +255,7 @@ func (m *Face) SetSubjectUID(subjUID string) (err error) {
 		Where("subj_src = ?", SrcAuto).
 		Where("subj_uid <> ?", m.SubjUID).
 		Where("marker_invalid = 0").
-		Updates(Values{"SubjUID": m.SubjUID, "MarkerReview": false}).Error; err != nil {
+		UpdateColumns(Values{"subj_uid": m.SubjUID, "marker_review": false}).Error; err != nil {
 		return err
 	}
 
@@ -271,9 +268,11 @@ func (m *Face) RefreshPhotos() error {
 		return fmt.Errorf("empty face id")
 	}
 
-	return UnscopedDb().Exec(`UPDATE photos SET checked_at = NULL WHERE id IN
-		(SELECT f.photo_id FROM files f JOIN ? m ON m.file_uid = f.file_uid WHERE m.face_id = ? GROUP BY f.photo_id)`,
-		gorm.Expr(Marker{}.TableName()), m.ID).Error
+	update := fmt.Sprintf(
+		"UPDATE photos SET checked_at = NULL WHERE id IN (SELECT f.photo_id FROM files f JOIN %s m ON m.file_uid = f.file_uid WHERE m.face_id = ?)",
+		Marker{}.TableName())
+
+	return UnscopedDb().Exec(update, m.ID).Error
 }
 
 // Hide hides the face by default.
@@ -304,6 +303,13 @@ func (m *Face) Create() error {
 
 // Delete removes the face from the database.
 func (m *Face) Delete() error {
+	// Remove face id from markers before deleting.
+	if err := Db().Model(&Marker{}).
+		Where("face_id = ?", m.ID).
+		UpdateColumns(Values{"face_id": "", "face_dist": -1}).Error; err != nil {
+		return err
+	}
+
 	return Db().Delete(m).Error
 }
 
@@ -349,4 +355,21 @@ func FindFace(id string) *Face {
 	}
 
 	return &f
+}
+
+// ValidFaceCount counts the number of valid face markers for a file uid.
+func ValidFaceCount(fileUID string) (c int) {
+	if !rnd.IsPPID(fileUID, 'f') {
+		return
+	}
+
+	if err := Db().Model(Marker{}).
+		Where("file_uid = ? AND marker_type = ?", fileUID, MarkerFace).
+		Where("marker_invalid = 0").
+		Count(&c).Error; err != nil {
+		log.Errorf("file: %s (count faces)", err)
+		return 0
+	} else {
+		return c
+	}
 }
